+++ cohere/responses/custom_model.py
@@ -2,10 +2,12 @@ from dataclasses import dataclass
 from datetime import datetime
 from typing import Any, Dict, Optional
 
+from cohere.utils import JobWithStatus
+
 try:
-    from typing import Literal
+    from typing import Literal, TypedDict
 except ImportError:
-    from typing_extensions import Literal
+    from typing_extensions import Literal, TypedDict
 
 from cohere.responses.base import CohereObject
 
@@ -26,25 +28,57 @@ CUSTOM_MODEL_STATUS = Literal[
     "QUEUED",
 ]
 
-INTERNAL_CUSTOM_MODEL_TYPE = Literal[
-    "GENERATIVE",
-    "CONTRASTIVE",
-    "CLASSIFICATION",
-]
-CUSTOM_MODEL_TYPE = Literal["GENERATIVE", "EMBED", "CLASSIFY"]
+INTERNAL_CUSTOM_MODEL_TYPE = Literal["GENERATIVE", "CLASSIFICATION", "RERANK"]
+CUSTOM_MODEL_TYPE = Literal["GENERATIVE", "CLASSIFY", "RERANK"]
 CUSTOM_MODEL_PRODUCT_MAPPING: Dict[CUSTOM_MODEL_TYPE, INTERNAL_CUSTOM_MODEL_TYPE] = {
     "GENERATIVE": "GENERATIVE",
-    "EMBED": "CONTRASTIVE",
     "CLASSIFY": "CLASSIFICATION",
+    "RERANK": "RERANK",
 }
 REVERSE_CUSTOM_MODEL_PRODUCT_MAPPING: Dict[INTERNAL_CUSTOM_MODEL_TYPE, CUSTOM_MODEL_TYPE] = {
     v: k for k, v in CUSTOM_MODEL_PRODUCT_MAPPING.items()
 }
 
 
-class CustomModel(CohereObject):
+@dataclass
+class HyperParameters:
+    early_stopping_patience: int
+    early_stopping_threshold: float
+    train_batch_size: int
+    train_steps: int
+    learning_rate: float
+
+    @staticmethod
+    def from_response(response: Optional[dict]) -> "HyperParameters":
+        return HyperParameters(
+            early_stopping_patience=response["earlyStoppingPatience"],
+            early_stopping_threshold=response["earlyStoppingThreshold"],
+            train_batch_size=response["trainBatchSize"],
+            train_steps=response["trainSteps"],
+            learning_rate=response["learningRate"],
+        )
+
+
+class HyperParametersInput(TypedDict):
+    """
+    early_stopping_patience: int (default=6, min=0, max=10)
+    early_stopping_threshold: float (default=0.01, min=0, max=0.1)
+    train_batch_size: int (default=16, min=2, max=16)
+    train_steps: int (default=2500, min=100, max=20000)
+    learning_rate: float (default=0.01, min=0.000005, max=0.1)
+    """
+
+    early_stopping_patience: int
+    early_stopping_threshold: float
+    train_batch_size: int
+    train_steps: int
+    learning_rate: float
+
+
+class BaseCustomModel(CohereObject, JobWithStatus):
     def __init__(
         self,
+        wait_fn,
         id: str,
         name: str,
         status: CUSTOM_MODEL_STATUS,
@@ -52,6 +86,7 @@ class CustomModel(CohereObject):
         created_at: datetime,
         completed_at: Optional[datetime],
         model_id: Optional[str] = None,
+        hyperparameters: Optional[HyperParameters] = None,
     ) -> None:
         super().__init__()
         self.id = id
@@ -61,10 +96,13 @@ class CustomModel(CohereObject):
         self.created_at = created_at
         self.completed_at = completed_at
         self.model_id = model_id
+        self.hyperparameters = hyperparameters
+        self._wait_fn = wait_fn
 
     @classmethod
-    def from_dict(cls, data: Dict[str, Any]) -> "CustomModel":
+    def from_dict(cls, data: Dict[str, Any], wait_fn) -> "BaseCustomModel":
         return cls(
+            wait_fn=wait_fn,
             id=data["id"],
             name=data["name"],
             status=data["status"],
@@ -72,7 +110,59 @@ class CustomModel(CohereObject):
             created_at=_parse_date(data["created_at"]),
             completed_at=_parse_date(data["completed_at"]) if "completed_at" in data else None,
             model_id=data["model"]["route"] if "model" in data else None,
+            hyperparameters=HyperParameters.from_response(data["settings"]["hyperparameters"])
+            if data["settings"]["hyperparameters"]
+            else None,
         )
+
+    def has_terminal_status(self) -> bool:
+        return self.status == "READY"
+
+
+class CustomModel(BaseCustomModel):
+    def wait(
+        self,
+        timeout: Optional[float] = None,
+        interval: float = 60,
+    ) -> "CustomModel":
+        """Wait for custom model job completion.
+
+        Args:
+            timeout (Optional[float], optional): Wait timeout in seconds, if None - there is no limit to the wait time.
+                Defaults to None.
+            interval (float, optional): Wait poll interval in seconds. Defaults to 60.
+
+        Raises:
+            TimeoutError: wait timed out
+
+        Returns:
+            CustomModel: custom model.
+        """
+
+        return self._wait_fn(custom_model_id=self.id, timeout=timeout, interval=interval)
+
+
+class AsyncCustomModel(BaseCustomModel):
+    async def wait(
+        self,
+        timeout: Optional[float] = None,
+        interval: float = 60,
+    ) -> "CustomModel":
+        """Wait for custom model job completion.
+
+        Args:
+            timeout (Optional[float], optional): Wait timeout in seconds, if None - there is no limit to the wait time.
+                Defaults to None.
+            interval (float, optional): Wait poll interval in seconds. Defaults to 60.
+
+        Raises:
+            TimeoutError: wait timed out
+
+        Returns:
+            CustomModel: custom model.
+        """
+
+        return await self._wait_fn(custom_model_id=self.id, timeout=timeout, interval=interval)
 
 
 @dataclass